pyproject.toml
[tool.poetry]
name = "ben-wiki"
version = "0.1.0"
description = "A wiki using mkdocs and mkdocs jupyter meant to serve as a portfolio and blog for my software engineering endeavors."
authors = ["Ben Hill <benjamin.hill28@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
mkdocs = "^1.3.0"
mkdocs-jupyter = "^0.20.1"
jupyter = "^1.0.0"
matplotlib = "^3.5.1"
mkdocs-windmill-dark = "^0.2.0"
mkdocs-material = "^8.2.8"
flake8 = "^4.0.1"
pytest = "^7.1.1"
numpy = "^1.22.3"
pandas = "^1.4.2"
mysql-connector-python = "^8.0.29"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
